requirements.txt
praw==7.7.1
python-dotenv==1.0.1
requests==2.32.3
google-generativeai==0.7.2
tenacity==9.0.0
pytz==2024.1
packaging==24.1
# v2 additions
tweepy==4.14.0
pytrends==4.9.2
pillow==10.4.0
pytesseract==0.3.13
opencv-python==4.10.0.84
ffmpeg-python==0.2.0
yt-dlp==2024.08.06
boto3==1.34.162
